added minimising algorithm and tests
implemented an entry point in main.py where a regex can be transfored into a minimised dfa
and where there can be created diagrams for the nfa, the dfa and the minimal dfa.

added run_tests.py to run the json file provided for the resulted dfa and the minimal dfa.

used Hopcroft's dfa minimisation algorithm in minimise.py